Fixed Windows version. New setting: windowsScript.
The Windows version did not work due to the sourced script
(Tclssg) not knowing what to do with the event loop started by
the Starpack on Windows. Rather than modify the script I
introduced an option to specify an anonymous function with which
to override the default behavior on Windows (which could also
be reused for other platforms without a real console). The default
setting for this option shows the Tk console window then sources
and runs the Tclssg in a new thread. It redirects all calls to
[puts] to the Tk console.