feat: build multiple starpacks in one go
Replace the option "targetTclkit" with "targetTclkits", a list of paths.
Rename the option "checkout" "revision".  Do not process the revision
"%AUTO%" (use, e.g., "HEAD" instead).

Put the application in the subdirectory app/ of the VFS to avoid path
conflicts.

Simplify the main and the Windows script.  Make them not lambdas.  Set
::argv0 to the path to fileToSource in the main script.

Add usage and the ability override options with command line arguments
in build.tcl.

Require Tcl 8.6.

Update the copyright.

Bump the version to 0.7.0.